chore(shared-state): refactor sharedstate error types
